캠프 테스트

+++ src/main/java/com/jeju_campking/campking/camp/entity/Camp.java
@@ -21,7 +21,7 @@ import lombok.*;
 @Builder
 public class Camp {
 
-    int campNumber;
+    Long campNumber;
     String campAddress;
     String campDetail;
     String campTypeNormal;
